Group all apps specific code in main in own folder
